New: Add unit tests for v2/v3 manifest parsing
