+++ test/manifests/iiifv3.json
@@ -1,4 +1,4 @@
-export default {
+{
     "@context":[
         "http://www.w3.org/ns/anno.jsonld",
         "http://iiif.io/api/presentation/3/context.json"
